+++ packages/styles/src/packages/atomic/breadcrumbs/index.ts
@@ -1,0 +1,6 @@
+import * as styles from "./rarui-breadcrumbs.css";
+
+export const breadcrumbs = {
+  classnames: { ...styles },
+};
+// export type { CheckboxVariants } from "./rarui-breadcrumbs.types";