+++ packages/styles/src/packages/atomic/breadcrumbs/rarui-breadcrumbs.types.ts
@@ -1,0 +1,4 @@
+import { RecipeVariants } from "@vanilla-extract/recipes";
+import { styles } from "./rarui-breadcrumbs.css";
+
+export type BreadcrumbsVariants = RecipeVariants<typeof styles.breadcrumbs>;